Card Console Updates
Unless you want a lieutenant to personally allow more lieutenants aboard the station, this is a tweak which should happen.

Also, the SM now has their own personal department management. If I didn't fuck it up.

This is by no means a final thing and should not stand unless the commit is amended to show test success.